+++ pages/_app.tsx
@@ -1,10 +1,6 @@
 import '@/styles/globals.css';
 import { createBrowserSupabaseClient } from '@supabase/auth-helpers-nextjs';
-import {
-  SessionContextProvider,
-  Session,
-  useUser,
-} from '@supabase/auth-helpers-react';
+import { SessionContextProvider, Session } from '@supabase/auth-helpers-react';
 import { AppProps } from 'next/app';
 import { Source_Sans_3 } from '@next/font/google';
 import { ToastContainer } from 'react-toastify';
