+++ tsconfig.json
@@ -23,7 +23,8 @@
       "@/types/*": ["types/*"],
       "@/styles/*": ["styles/*"],
       "@/context/*": ["context/*"],
-      "@/public/*": ["public/*"]
+      "@/public/*": ["public/*"],
+      "@/hooks/*": ["hooks/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
